Setter opp kjøring fra docker
