+++ .github/workflows/deploy-naisjob.yml
@@ -42,4 +42,4 @@ jobs:
           CLUSTER: prod-gcp
           IMAGE: ${{ needs.docker-build-push.outputs.image }}
           APIKEY: ${{ secrets.NAIS_DEPLOY_APIKEY }}
-          RESOURCE: .nais.yaml
+          RESOURCE: .nais/prod.yaml
