+++ requirements.txt
@@ -1,7 +1,6 @@
-google-cloud-bigquery==3.12.0
 google-auth==2.23.2
-datastory==0.1.16
+google-cloud-bigquery==3.12.0
 pandas==2.1.1
 plotly==5.17.0
 jupyter==1.0.0
-black==23.9.1
+tabulate~=0.9.0
